Separate the four precision questions, and stop averaging a peak in silence
Measured three things before writing any of them down.

A float32 field of 300 +/- 0.001, ten million values, accumulated in
float32 has a mean of exactly 300.000000000000 against an exact
299.999999895342 - the variation the field was written to carry is
gone and the offset is printed to full width. Accumulated in float64
it is 160 times closer and costs nothing (E-143).

A stress concentration inside a body, element values 10/20/200/20/10
MPa, averaged onto its shared nodes has a maximum of 110 MPa against
200. Fifty-five per cent, a 90 MPa under-report. The same peak at an
end face gives 200 either way, so a check placed at a boundary reports
that averaging is harmless (E-144).

And the vendor documentation for the reference post-processor, which
publishes the spread as Nodal Difference and says a large one means
the mesh needs refining there (E-145).

Four precision questions were being treated as one: store what the
file gave, compute in double, show what the source supports, and hold
geometry in double regardless. XC-246 separates them.

INV-032 and XC-247: both extrema are produced, each says which it is,
and an averaged peak carries the spread it was averaged from. INV-033
forbids implying convergence the product cannot see.

One claim of mine was wrong and is corrected in place with the record
kept. E-143 first read the 0.0 difference of two float32 values as a
loss in the subtraction. It is not - both literals round to the same
float32 before anything is subtracted, and subtraction within a factor
of two is exact, checked over 100,000 pairs. What a near-equal
difference loses is significance: ten digits in, one out, and the
storage type says nothing about it. That is INV-034, which is new
because the wrong reading was corrected rather than deleted.

+++ src/engine/analysis/difference.py
@@ -133,7 +133,11 @@ def difference(
                 f"'{name}' の要素数が違います（{a.values.shape[0]} と {b.values.shape[0]}）。"
                 "識別子がないため位置で対応させるほかなく、長さが違えば対応がつきません"
             )
-        values = a.values - b.values
+        # Subtracted in float64 whatever the fields are stored in (INV-031, XC-246). Two float32
+        # values 1e-7 apart subtract to **exactly 0.0**, and a difference of zero is what an engineer
+        # reads as "these agree" - so the storage precision of the inputs would decide whether a real
+        # difference is reported at all (E-143).
+        values = a.values.astype(np.float64) - b.values.astype(np.float64)
         unmatched = 0
         method = Method.SHARED_MESH
 
@@ -169,7 +173,9 @@ def _by_identifier(
             # engineer reads as "these agree" (INV-011).
             unmatched += 1
             continue
-        values[index] = left_values[index] - right_values[found]
+        # float64 on both sides before the subtraction, for the reason in INV-031: the cancellation
+        # here is between two nearly equal numbers, which is where single precision returns zero.
+        values[index] = float(left_values[index]) - float(right_values[found])
     return values, unmatched
